Move download chart functionality into component
