--- src/lib/download-file.js
@@ -1,7 +1,0 @@
-export default function downloadFile(filename, dataURL) {
-  const anchor = document.createElement("a");
-  anchor.href = dataURL;
-  anchor.target = "_blank";
-  anchor.download = filename;
-  anchor.click();
-}
